Clean up main file by moving some things to headers.
